Remove functions.
