+++ src/tomato/tomato/__init__.py
@@ -1,5 +1,16 @@
 """
-Main module - executables for tomato.
+**tomato**: command line interface to the tomato daemon
+-------------------------------------------------------
+.. codeauthor:: 
+    Peter Kraus
+
+Module of functions to interact with tomato. Includes job management functions:
+
+- :func:`.submit` to submit a *job* to *queue*
+- :func:`.status` to query the status of tomato's *pipelines*, its *queue*, or a *job*
+- :func:`.cancel` to cancel a queued or kill a running *job*
+- :func:`.snapshot` to create an up-to-date FAIR data archive of a running *job*
+- :func:`.search` to find a ``jobid`` of a *job* from ``jobname``
 
 """
 import os
